+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Create app CR for per cluster app-operator instance.
+- Add `appfinalizer` resource to remove finalizers from workload cluster app CRs.
+
 ## [0.24.2] - 2021-02-25
 
 ### Changed