+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.7.0] - 2021-03-15
+
 ## [3.6.1] - 2021-02-23
 
 ### Removed
@@ -218,7 +220,8 @@ be edited by users.
 
 
 
-[Unreleased]: https://github.com/giantswarm/cluster-operator/compare/v3.6.1...HEAD
+[Unreleased]: https://github.com/giantswarm/cluster-operator/compare/v3.7.0...HEAD
+[3.7.0]: https://github.com/giantswarm/cluster-operator/compare/v3.6.1...v3.7.0
 [3.6.1]: https://github.com/giantswarm/cluster-operator/compare/v3.6.0...v3.6.1
 [3.6.0]: https://github.com/giantswarm/cluster-operator/compare/v3.5.1...v3.6.0
 [3.5.1]: https://github.com/giantswarm/cluster-operator/compare/v3.5.0...v3.5.1